Break up foscam.js into separate files based on functionality
