+++ .jsdoc.json
@@ -3,7 +3,7 @@
     "dictionaries": ["jsdoc"]
   },
   "source": {
-    "include": ["foscam.js", "package.json", "README.md"],
+    "include": ["lib", "package.json", "README.md"],
     "includePattern": ".js$",
     "excludePattern": "(node_modules/|docs)"
   },
