+++ index.js
@@ -1,4 +1,4 @@
-var Foscam = require('./foscam');
+var Foscam = require('./lib/index');
 var config = require('./configs');
 
 var api = new Foscam(config.username, config.password, config.host);
