+++ lib/index.js
@@ -1,0 +1,7 @@
+var Foscam = require('./Foscam');
+require('./AV');
+require('./User');
+require('./PTZ');
+require('./Network');
+
+module.exports = Foscam;
